configure doh as module
