+++ doh/__main__.py
@@ -1,0 +1,5 @@
+import sys
+from doh import doh
+
+if __name__ == "__main__":
+    sys.exit(doh.launch())
