+++ requirements.txt
@@ -15,6 +15,7 @@ pytest==8.3.3
 python-slugify==8.0.4
 PyYAML==6.0.2
 requests==2.32.3
+setuptools==75.6.0
 snowballstemmer==2.2.0
 Sphinx==8.1.3
 sphinx-rtd-theme==3.0.2
